Store isolated-nonbonded per-group energies in mixed precision
The IsolatedNonbondedForce per-group energy breakdown buffer was float, narrowing
the accumulation to fp32 even though the main energy is already deterministic
double (fixed-point). Make groupEnergies mixed* in the kernel (drop the (float)
cast on the real pairEnergy) and allocate/read it back at the context precision on
the host (local mixed helpers; groupEnergiesHost -> vector<double>; getGroupEnergy
already returns double). The deterministic fixed-point main-energy accumulator is
unchanged.

test_particle_groups adds isolated-NB per-group cases: the breakdown is finite and
sums to the total in single/mixed/double.

+++ platforms/cuda/src/CudaIsolatedNonbondedKernels.h
@@ -83,7 +83,7 @@ private:
     OpenMM::CudaArray groupScalingFactorsBuffer;  // [numGroups] - per-group scaling
 
     // Host-side cached results
-    mutable std::vector<float> groupEnergiesHost;
+    mutable std::vector<double> groupEnergiesHost;
 
     std::vector<int> h_particleIndices;  // Host copy for updates
